+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Removed `World.Batch` for entity batch creation, use `Builder` instead (#239)
 * Rework of generic entity creation API, use `MapX.New`, `MapX.NewWith`, `MapX.NewBatch` and `MapX.NewQuery` (#239, #252)
+* Stats object `WorldStats` etc. adapted for new structure of archetypes nested in nodes (#258)
 
 ### Features